refactor: move `rpc` to a subdirectory, start of generalized logic
